Move pkg/teams to pkg/input/teams
